src/main/persistence/JsonWriter.java
package src.main.persistence;

import org.json.JSONObject;
import src.main.model.company.TradersMap;

import java.io.File;
import java.io.FileNotFoundException;
import java.io.PrintWriter;

// Reference: JsonSerializationDemo
// Represents a writer that writes JSON representation of a TradersMap to file
public class JsonWriter {
    private static final int TAB = 4;
    private PrintWriter writer;
    private String destination;

    // EFFECTS: constructs writer to write to destination file
    public JsonWriter(String destination) {
        this.destination = destination;
    }

    // MODIFIES: this
    // EFFECTS: opens writer; throws FileNotFoundException if destination file cannot
    //          be opened for writing
    public void open() throws FileNotFoundException {
        writer = new PrintWriter(new File(destination));
    }

    // MODIFIES: this
    // EFFECTS: writes JSON representation of TradersMap to file
    public void write(TradersMap tm) {
        JSONObject json = tm.toJson();
        saveToFile(json.toString(TAB));
    }

    // MODIFIES: this
    // EFFECTS: closes writer
    public void close() {
        writer.close();
    }

    // MODIFIES: this
    // EFFECTS: writes string to file
    private void saveToFile(String json) {
        writer.print(json);
    }

}
